redefine client label, add host_name label
Formerly the client metrics `client` label was populated with the hostname provided by the controller. The hostname parameter is not reliably present.

This change populates the `client` label with the name value from the controller, and moves the hostname to a separate `hostname` label.

It is not 100% clear how Omada sets the name. A best guess is:

- If the name explicitly set a name in the Omada UI, that.
- If no name is explicitly set, and a hostname has been determined, use the hostname.
- If no name is explicitly set, and no hostname has been determined, us the mac address.