fix client signal level bug, add client signal-to-noise ratio (#88)
* go fmt (whitespace changes only)

* (bug) signal level is a percent, not dBm

* Add client signal to noise ratio

* fix missing client values

The HostName client field seems somewhat unreliable in practice. Fall back on the Name field if it is absent.

Maybe a fix for #22

* add an mdocs command to emit metrics docs

This should guarantee docs matching the code, and replace the defunct metric-markdown-table.sh script.

* refresh metrics descriptions

* replace metric-markdown-table.sh with mdocs command

Remaining buglet: the version and mdocs commands shouldn’t require host, username and password arguments.

* redefine client label, add host_name label

Formerly the client metrics `client` label was populated with the hostname provided by the controller. The hostname parameter is not reliably present.

This change populates the `client` label with the name value from the controller, and moves the hostname to a separate `hostname` label.

It is not 100% clear how Omada sets the name. A best guess is:

- If the name explicitly set a name in the Omada UI, that.
- If no name is explicitly set, and a hostname has been determined, use the hostname.
- If no name is explicitly set, and no hostname has been determined, us the mac address.